feat: add order persistence layer (entities, DTOs, repositories, service)
- Create JPA entities matching existing DB schema (Category, SubCategory, Customer, Seller, Product, Orders, Shipment, Payment, OrderItem, OrderStatus enum)
- Create DTOs for JSON deserialization of Pub/Sub order payloads
- Create Spring Data JPA repositories with upsert support (findByDocument, findByProductId)
- Create OrderPersistenceService with @Transactional persist logic respecting FK order
- Edit PubSubMessageHandler to deserialize and persist incoming messages
- Fix entity types (Integer vs Long) to match PostgreSQL SERIAL columns
- Set ddl-auto=none since schema is managed by init SQL
- Update PubSubMessageHandlerTest with mocked dependencies

+++ backend-v1/src/main/java/br/com/fatec/messaging/repository/OrderItemRepository.java
@@ -1,11 +1,9 @@
 package br.com.fatec.messaging.repository;
 
+import br.com.fatec.messaging.model.OrderItem;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import br.com.fatec.messaging.model.OrderItem;
-
 @Repository
-public interface OrderItemRepository extends JpaRepository<OrderItem, Long> {
-
+public interface OrderItemRepository extends JpaRepository<OrderItem, Integer> {
 }
