+++ backend-v1/src/main/java/br/com/fatec/messaging/repository/PaymentRepository.java
@@ -1,11 +1,9 @@
 package br.com.fatec.messaging.repository;
 
+import br.com.fatec.messaging.model.Payment;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import br.com.fatec.messaging.model.Payment;
-
 @Repository
-public interface PaymentRepository extends JpaRepository<Payment, Long> {
-
+public interface PaymentRepository extends JpaRepository<Payment, Integer> {
 }
