+++ backend-v1/src/main/java/br/com/fatec/messaging/repository/ShipmentRepository.java
@@ -1,11 +1,9 @@
 package br.com.fatec.messaging.repository;
 
+import br.com.fatec.messaging.model.Shipment;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import br.com.fatec.messaging.model.Shipment;
-
 @Repository
-public interface ShipmentRepository extends JpaRepository<Shipment, Long> {
-
+public interface ShipmentRepository extends JpaRepository<Shipment, Integer> {
 }
